Move DataContainer to a new file.
